refactor: Move LoginForm and RegisterForm to components/forms directory
